feat: completely replace user provided documentclass
To resolve bad UX when user copies and pastes non-dvisvgm compliant TikZ
example.  The downside of this change is that it prevents users from
customizing the documentclass.

This implements option 1 from
https://github.com/yuzutech/kroki/issues/1870#issuecomment-2822066909

NOTE: This change requires GNU sed.  This change is not compatible with
the BusyBox version of sed that Alpine defaults to.